backend/requirements.txt
# Web framework
fastapi
uvicorn[standard]

# Environment variables
python-dotenv

# Data validation
pydantic

# Database
psycopg2-binary

# Audio & video
yt-dlp
ffmpeg-python

# Speech-to-text (offline)
openai-whisper
faster-whisper

# NLP / Summarization
transformers
torch
sentencepiece
sumy
nltk
sentence-transformers
numpy

# LLM / AI
groq
faiss-cpu

# JWTs
PyJWT

# Utilities
regex
